Changed QuantityIncluded from Int to Double

+++ CheddarGetterDataClasses.cs
@@ -63,7 +63,7 @@ namespace Confer.CheddarGetter
         public Guid ID { get; set; }
         public string Code { get; set; }
         public string Name { get; set; }
-        public int QuantityIncluded { get; set; }
+        public double QuantityIncluded { get; set; }
         public bool IsPeriodic { get; set; }
         public float OverageAmount { get; set; }
         public DateTime CreatedDateTime { get; set; }
